fixed an issue with route progress bar margin left and created a hook for progress counter
