--- components/Progress/Bar.types.ts
@@ -1,9 +1,0 @@
-export interface IProps {
-  animationDuration: number;
-  progress: number;
-}
-
-export interface BarStyleProps {
-  ml: string;
-  tr: string;
-}
