--- components/Progress/Container.types.ts
@@ -1,10 +1,0 @@
-export interface IProps {
-  animationDuration: number;
-  isFinished: boolean;
-  children: any;
-}
-
-export interface ContainerStyleProps {
-  isF: boolean;
-  tr: string;
-}
