+++ components/Progress/Progress.types.ts
@@ -1,3 +1,13 @@
 export interface IProps {
   isAnimating: boolean;
 }
+
+export interface ContainerStyleProps {
+  isF: boolean;
+  tr: string;
+}
+
+export interface BarStyleProps {
+  ml: string;
+  tr: string;
+}
