+++ hooks/index.ts
@@ -1,0 +1,3 @@
+import useAnimatedRouteChange from "./useAnimatedRouteChange";
+
+export { useAnimatedRouteChange };
