+++ pages/_app.tsx
@@ -1,30 +1,10 @@
-import { useEffect } from "react";
-import { useRouter } from "next/router";
-import type { AppProps } from "next/app";
 import GlobalStyles from "../styles/Global";
 import Progress from "../components/Progress";
-import { useProgressStore } from "../store";
+import { useAnimatedRouteChange } from "../hooks";
+import type { AppProps } from "next/app";
 
 function MyApp({ Component, pageProps }: AppProps) {
-  const router = useRouter();
-  const setIsAnimating = useProgressStore((state: any) => state.setIsAnimating);
-  const isAnimating = useProgressStore((state: any) => state.isAnimating);
-
-  useEffect(() => {
-    const handleStartAnimating = () => setIsAnimating(true);
-    const handleStopAnimating = () => setIsAnimating(false);
-
-    router.events.on("routeChangeStart", handleStartAnimating);
-    router.events.on("routeChangeComplete", handleStopAnimating);
-    router.events.on("routeChangeError", handleStopAnimating);
-
-    return () => {
-      router.events.off("routeChangeStart", handleStartAnimating);
-      router.events.off("routeChangeComplete", handleStopAnimating);
-      router.events.off("routeChangeError", handleStopAnimating);
-    };
-  }, [router, setIsAnimating]);
-
+  const { isAnimating } = useAnimatedRouteChange();
   return (
     <>
       <Progress isAnimating={isAnimating} />
@@ -33,5 +13,4 @@ function MyApp({ Component, pageProps }: AppProps) {
     </>
   );
 }
-
 export default MyApp;
